Adding comments to about half of the scripts

+++ scripts/deletepicture.php
@@ -4,13 +4,13 @@
 
   require_once "./dbconnect.php";
 
-  if(!isset($_SESSION["loggedIn"]["role_ID"]) || $_SESSION["loggedIn"]["role_ID"] == 1)
+  if(!isset($_SESSION["loggedIn"]["role_ID"]) || $_SESSION["loggedIn"]["role_ID"] == 1) //Is the correct user type accessing this script?
   {
     header("location: ../pages/index.php");
     exit();
   }
 
-  if(!isset($_GET["picid"]))
+  if(!isset($_GET["picid"]))  //Check if the Picture ID has been passed
   {
     header("location: ../pages/index.php");
     exit();
@@ -19,7 +19,7 @@
   $sqlsel = "SELECT id, picture_link FROM pictures WHERE id=$_GET[picid]";
   $resultsel = $conn->query($sqlsel);
 
-  if ($resultsel->num_rows != 1)
+  if ($resultsel->num_rows != 1)  //Check if Picture ID exists
   {
     $_SESSION["error"] = "Coś poszło nie tak. Czy zdjęcie nie zostało już usunięte?";
     echo "<script>history.back()</script>";
@@ -33,14 +33,14 @@
 
   $sqlpic = "DELETE FROM pictures WHERE id = $_GET[picid]";
   $conn->query($sqlpic);
-  if ($conn->affected_rows == 1)
+  if ($conn->affected_rows == 1)  //Check if SQL executed properly and affected 1 row
   {
-    $_SESSION["success"] = "Zdjęcie zostało usunięte!";
+    $_SESSION["success"] = "Zdjęcie zostało usunięte!"; //Announce success and return to the previous page
     unlink(realpath($prod["picture_link"]));
     echo "<script>history.back()</script>";
     exit();
   }
-  else
+  else  //Throw an error
   {
     $_SESSION["error"] = "Coś poszło nie tak. Skontaktuj się z administratorem.";
     echo "<script>history.back()</script>";
